template/tools/orchestrator/requirements.txt
flask==3.0.0
redis==5.0.1
pyyaml==6.0.1
requests==2.31.0
anthropic==0.18.1
gitpython==3.1.40
